+++ packages/expo-constants/CHANGELOG.md
@@ -4,9 +4,14 @@
 
 ### 🛠 Breaking changes
 
+- Use an explicit development or production mode when loading native app config and env files. ([#49511](https://github.com/expo/expo/pull/49511) by [@ramonclaudio](https://github.com/ramonclaudio))
+
 ### 🎉 New features
 
 ### 🐛 Bug fixes
+
+- [Android] Generate native app config using the Constants library's build type. ([#49511](https://github.com/expo/expo/pull/49511) by [@ramonclaudio](https://github.com/ramonclaudio))
+- [iOS] Use production config by default when Updates native debugging embeds production JavaScript. ([#49511](https://github.com/expo/expo/pull/49511) by [@ramonclaudio](https://github.com/ramonclaudio))
 
 ### 💡 Others